CR Feedback (Refactor via TransformOperation.transform() method).
